Add function to get the blacklisted verifiers from service manager

+++ contracts/src/core/AlignedLayerServiceManagerStorage.sol
@@ -12,7 +12,7 @@ abstract contract AlignedLayerServiceManagerStorage {
     mapping(bytes32 => BatchState) public batchesState;
 
     // stores blacklisted verifiers
-    uint256 public blacklistedVerifiers;
+    uint64 public blacklistedVerifiers;
 
     // Storage for batchers balances. Used by aggregator to pay for respondToTask
     mapping(address => uint256) public batchersBalances;
